basic fixture movement w/ conftest
