+++ tests/test_score.py
@@ -4,9 +4,22 @@ from lxml import etree
 
 from lejaren.notation import Note, Part, Score, Chord
 
-
+@pytest.mark.skip
 def test_xml_valid():
-    pass
+
+    schema_doc = etree.parse("musicxml-4.0/schema/musicxml.xsd")
+    schema = etree.XMLSchema(schema_doc)
+
+    xml_doc = etree.parse("data.xml")
+
+    if schema.validate(xml_doc):
+        print("Success: XML is fully valid.")
+    else:
+        print("Error: XML configuration violates schema definitions.")
+        # Log precise errors
+        for error in schema.error_log:
+            print(f"Line {error.line}: {error.message}")
+
 
 
 def test_whole_note_chord_in_XML():
